Implemented bar chart & some testing code
The bar chart shows how many times the button was pressed for each day of the past 7 days. However, due to some timezone issues, the dates are not accurate to the database table. Everything else is just for my own personal testing and will be removed once everything is working properly