File upload logic
Partial commit of file upload logic
This code establishes a connection between the file dropzone and the server.
Have tested and confirmed working.
Application is able to accept a .txt file and send it to the server.

- FileDropzone sends .txt as formdata to server via an Axios post request
- Server receives file using multer, a node JS middleware for handling form-data
- File content is logged to the console

Next phase will involve analysing the .txt for data and then inserting this into the database.